VWO: no more unusedMap

+++ src/main/java/nl/v4you/jafs/internal/JafsInode.java
@@ -301,7 +301,7 @@ public class JafsInode {
 		return Util.arrayToInt(bb1, 0);
 	}
 
-	void writeInt(int i) throws JafsException, IOException {
+	void writeInt(long i) throws JafsException, IOException {
 		Util.intToArray(bb1, 0, i);
 		writeBytes(bb1, 0, 4);
 	}
